apps/api/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
pydantic[email]
supabase==2.6.0
httpx==0.27.2
typing-extensions==4.12.2
loguru==0.7.2
aiofiles==23.2.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4

